Update Python output to Dictionary instead of Array
Before it would always return an Array, the first element could be the
expected response or an error message.

Now returns a Dictionary with two keys "error" and "value". The first
contains a boolean to warn if is an error or not. The second is the same
as before, an array with as value that were suppose to return.

When the error flag is true, the value will be the error message, same
as before.