Create PythonPager, may contain breaking changes
PythonPager is the middle man between PythonWatcher scene and Python scene.
It will deal with page logic. For example, now it detects if is the last
page and doesn't request more pages after that.

The breaking change came with the fact that PythonPage understand that
the output is a JSON and get the error message inside it.

With this change, PythonWatcher doesn't give you the content String
anymore. Now it gives you the GenericResult containing any error and the
list of results from the Python operation.

Example
Before the PythonWatcher woudl return to you something like:
"""
{
  "error": 0,
  "result": [
    {"example": "foobar"}
  ]
}
"""

Now it does return you a GenericResult scene:
GenericResult.error = 0
GenericResult.error_string = ""
GenericResult.result = [
  {"example": "foobar"}
]

This means that anyone that used PythonWatcher my fail if i don't update
what it does when receive the output.